+++ includes/widgets/widget-browser/ax-artifact-repository-link-widget/ax-artifact-repository-link-widget.js
@@ -17,7 +17,6 @@ define( [
    Controller.$inject = [ '$scope' ];
 
    function Controller( $scope ) {
-
       $scope.resources = {
          display: null
       };