+++ includes/widgets/widget-browser/ax-artifact-repository-link-widget/spec/spec_runner.js
@@ -1,14 +1,16 @@
 /**
  * Copyright 2015 aixigo AG
  * Released under the MIT license.
- * http://laxarjs.org
+ * http://www.laxarjs.org
  */
 ( function( global ) {
    'use strict';
    global.laxarSpec = {
-      title: 'AxArtifactRepositoryLinkWidget Specification',
+      title: 'ax-artifact-repository-link-widget Specification',
       tests: [
          'ax-artifact-repository-link-widget.spec'
-      ]
+      ],
+      testRunner: 'laxar-mocks',
+      jasmineMajorVersion: 2
    };
 } )( this );