+++ includes/widgets/widget-browser/ax-widget-browser-widget/ax-widget-browser-widget.js
@@ -58,6 +58,7 @@ define( [
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       function publishSelectedWidget() {
+
          if( !model.selectedWidgetName ) {
             return;
          }
@@ -224,16 +225,13 @@ define( [
 
       function createList() {
          model.list = resources.list.widgets.map( function( widget ) {
-            var nameParts = widget.name.split( /[_-]/g );
-            nameParts = nameParts.map( function( word ) {
-               return word.substring( 0, 1 ).toUpperCase() + word.substring( 1 );
-            } );
-            var name = nameParts.join( '' );
+            var name = widget.name;
             var placeParameters = {};
             placeParameters[ $scope.features.select.parameter ] =  name;
             $scope.features.select.parameterList.forEach( function( parameter ) {
                placeParameters[ parameter ] =  '';
             } );
+            var lol = ax.object.deepClone(flowService);
             return {
                name: name,
                href: flowService.constructAbsoluteUrl( '_self', placeParameters ),