+++ includes/widgets/widget-browser/ax-widget-browser-widget/spec/ax-widget-browser-widget.spec.js
@@ -4,140 +4,97 @@
  * http://laxarjs.org/license
  */
 define( [
-   '../ax-widget-browser-widget',
-   'laxar/laxar_testing',
+   'json!../widget.json',
+   'laxar-mocks',
    'angular-mocks',
-   'jquery',
-   './fixtures',
-   'text!laxar-path-widgets/widget-browser/ax-widget-browser-widget/default.theme/ax-widget-browser-widget.html',
-   'json!laxar-path-widgets/widget-browser/ax-widget-browser-widget/widget.json',
-   'json!laxar-path-widgets/widget-browser/ax-widget-browser-widget/bower.json'
-], function( widgetModule, ax, ngMocks, $, specData, widgetMarkup, widgetInformation, widgetBower ) {
+   './data/fixtures',
+   'json!../widget.json',
+   'json!../spec/data/bower.json'
+], function( descriptor, axMocks, ngMocks, data, widgetInformation, widgetBower ) {
    'use strict';
 
-   describe( 'An AxWidgetBrowserWidget', function() {
-      var testBed;
-      var $widget;
+   describe( 'An ax-widget-browser-widget', function() {
       var $httpBackend;
+      var $provide;
+      var widgetEventBus;
+      var widgetScope;
+      var testEventBus;
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-      function setupHttpMockData( $injector ) {
-         $httpBackend = $injector.get( '$httpBackend' );
-         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json' )
-            .respond( 200, widgetInformation );
-         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/bower.json' )
-            .respond( 200, widgetBower );
-         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/SPECIFICATION.md' )
-            .respond( 200, {} );
-         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/README.md' )
-            .respond( 404, { value: 'Not Found' } );
-         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/DOCUMENTATION.md' )
-            .respond( 404, { value: 'Not Found' } );
-      }
+      function createSetup( widgetConfiguration, constructAbsoluteUrl ) {
 
-      ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      function setupHttpMockData404( $injector ) {
-         $httpBackend = $injector.get( '$httpBackend' );
-         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json' )
-            .respond( 404, { value: 'Not Found' } );
-         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/bower.json' )
-            .respond( 404, { value: 'Not Found' } );
-         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/SPECIFICATION.md' )
-            .respond( 404, { value: 'Not Found' } );
-         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/README.md' )
-            .respond( 404, { value: 'Not Found' } );
-         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/DOCUMENTATION.md' )
-            .respond( 404, { value: 'Not Found' } );
-      }
-      ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      function constructAbsoluteUrl( place, optionalParameters ) {
-         return 'http://localhost:8000/index.html#/widgetBrowser/' +
-                optionalParameters[ testBed.scope.features.select.parameter ];
-      }
-      ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      function constructAbsoluteUrlWithAdditionalParameters( place, optionalParameters ) {
-         var additionalParameters = '';
-         if( optionalParameters.anchor === undefined ) {
-            additionalParameters = additionalParameters + '/references';
-         }
-         if( optionalParameters.userId === undefined ) {
-            additionalParameters = additionalParameters + '/03804';
-         }
-         if( optionalParameters.processId === undefined ) {
-            additionalParameters = additionalParameters + '/73929';
-         }
-         return 'http://localhost:8000/index.html#/widgetBrowser/' +
-                optionalParameters[ testBed.scope.features.select.parameter ] + additionalParameters;
-      }
-
-      ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      function setup( features, onBeforeControllerCreationFunction, constructAbsoluteUrl  ) {
-         testBed = ax.testing.portalMocksAngular.createControllerTestBed( 'widget-browser/ax-widget-browser-widget' );
-         testBed.useWidgetJson();
-         testBed.injections = {
-            axFlowService: {
-               constructAbsoluteUrl: constructAbsoluteUrl
-            }
-         };
-         testBed.featuresMock = features;
-         testBed.setup( {
-            onBeforeControllerCreation: onBeforeControllerCreationFunction
+         beforeEach( function() {
+            ngMocks.module( function( _$provide_ ) {
+               $provide = _$provide_;
+            } );
          } );
-         ngMocks.inject( function( $compile ) {
-            $( '#container' ).remove();
-            $widget = $( '<div id="container"></div>' ).html( widgetMarkup );
-            $compile( $widget )( testBed.scope );
-            $widget.appendTo( 'body' );
+
+         beforeEach( axMocks.createSetupForWidget( descriptor ) );
+
+         beforeEach( function() {
+            $provide.value( 'axFlowService', {
+               constructAbsoluteUrl: constructAbsoluteUrl
+            } );
+            axMocks.widget.configure( widgetConfiguration );
+         } );
+
+         beforeEach( axMocks.widget.load );
+
+         beforeEach( function() {
+            widgetScope = axMocks.widget.$scope;
+            widgetEventBus = axMocks.widget.axEventBus;
+            testEventBus = axMocks.eventBus;
+            ngMocks.inject( function( $injector ) {
+               $httpBackend = $injector.get( '$httpBackend' );
+            } );
          } );
       }
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       afterEach( function() {
-         testBed.tearDown();
+         axMocks.tearDown();
       } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       describe( 'with a configured list resource', function() {
 
-         beforeEach( function() {
-            var features = {
-               list: {
-                  resource: 'listResource'
-               },
-               select: {
-                  parameter: 'widgetId'
-               }
+         var widgetConfiguration = {
+            list: {
+               resource: 'listResource'
+            },
+            select: {
+               parameter: 'widgetId'
+            }
+         };
 
-            };
-            setup( features, function() {}, constructAbsoluteUrl );
-            testBed.eventBusMock.publish( 'didReplace.listResource', {
-               resource: 'listResource',
-               data: specData.listResource
-            } );
-            jasmine.Clock.tick( 0 );
+         createSetup( widgetConfiguration, constructAbsoluteUrl );
+
+         beforeEach( function() {
+            axMocks.triggerStartupEvents();
+            testEventBus.publish( 'didReplace.listResource', {
+                     resource: 'listResource',
+                     data: data.listResource
+                  } );
+            testEventBus.flush();
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'acts as a slave of the list resource (R1.1)', function() {
-            expect( testBed.scope.eventBus.subscribe )
+            expect( widgetEventBus.subscribe )
                .toHaveBeenCalledWith( 'didReplace.listResource', jasmine.any( Function ) );
-            expect( testBed.scope.eventBus.subscribe )
+            expect( widgetEventBus.subscribe )
                .toHaveBeenCalledWith( 'didUpdate.listResource', jasmine.any( Function ) );
-            expect( testBed.scope.resources.list ).toEqual( specData.listResource );
+            expect( widgetScope.resources.list ).toEqual( data.listResource );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'displays a selectable list with all widgets and activities (R1.2)', function() {
-            expect( testBed.scope.model.list ).toEqual( specData.browserList );
+            expect( widgetScope.model.list ).toEqual( data.browserList );
          } );
 
       } );
@@ -147,58 +104,57 @@ define( [
       describe( 'after the user has made a selection', function() {
 
          describe( 'and bookmarked the site and entered the site through the bookmark', function() {
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  parameterList: [ 'anchor', 'userId' ],
+                  information: {
+                     resource: 'information'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrlWithAdditionalParameters );
 
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     parameterList: [ 'anchor', 'userId' ],
-                     information: {
-                        resource: 'information'
-                     }
-                  }
-               };
-               setup( features, function( $injector ) {
-                  $httpBackend = $injector.get( '$httpBackend' );
-                  $httpBackend.when( 'GET', /.*/ )
-                     .respond( 200, widgetInformation );
-               }, constructAbsoluteUrlWithAdditionalParameters );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               $httpBackend.when( 'GET', /.*/ ) .respond( 200, widgetInformation );
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'selects the widget in relation to the place parameter (R2.1)', function() {
-               expect( testBed.scope.model.selectedWidgetName ).toEqual( null );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               expect( widgetScope.model.selectedWidgetName ).toEqual( null );
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
-               expect( testBed.scope.model.selectedWidgetName ).toEqual( 'AxWidgetBrowserWidget' );
+               testEventBus.flush();
+               expect( widgetScope.model.selectedWidgetName ).toEqual( 'ax-widget-browser-widget' );
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'the links in the widget list only have the not excluded parameters (R2.2)', function() {
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget',
+                     widgetId: 'ax-widget-browser-widget',
                      anchor: 'references',
                      userId: '03804',
                      processId: '73929'
                   }
                } );
-               jasmine.Clock.tick( 0 );
-               expect( testBed.scope.model.list ).toEqual( specData.browserListWithParameters );
+               testEventBus.flush();
+               expect( widgetScope.model.list ).toEqual( data.browserListWithParameters );
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
@@ -206,27 +162,25 @@ define( [
             describe( 'and with a new resource list without the bookmarked widget and with a configured information resource', function() {
 
                beforeEach( function() {
-                  testBed.eventBusMock.publish( 'didReplace.listResource', {
+                  testEventBus.publish( 'didReplace.listResource', {
                      resource: 'listResource',
-                     data: specData.smallerListResource
+                     data: data.smallerListResource
                   } );
-                  jasmine.Clock.tick( 0 );
-                  testBed.eventBusMock.publish( 'didNavigate._self', {
+                  testEventBus.flush();
+                  testEventBus.publish( 'didNavigate._self', {
                      data: {
-                        widgetId: 'AxWidgetBrowserWidget'
+                        widgetId: 'ax-widget-browser-widget'
                      }
                   } );
-                  jasmine.Clock.tick( 0 );
+                  testEventBus.flush();
                } );
 
                ///////////////////////////////////////////////////////////////////////////////////////////////
 
-               it( 'doesn\'t try to get any widget information data and doesn\'t publish any resource (R2.3)', function() {
+               it( 'doesn\'t try to get any widget information data and replaces the resource with an empty object (R2.3)', function() {
                   $httpBackend.verifyNoOutstandingRequest();
-                  expect( testBed.scope.eventBus.publish )
-                     .not.toHaveBeenCalledWith( jasmine.any( String ), jasmine.any( Object ), jasmine.any( Object ) );
-                  expect( testBed.scope.eventBus.publish )
-                     .not.toHaveBeenCalledWith( jasmine.any( String ), jasmine.any( Object ) );
+                  expect( widgetEventBus.publish )
+                     .toHaveBeenCalledWith( 'didReplace.information', { resource: 'information', data: null }, { deliverToSender: false } );
                } );
             } );
          } );
@@ -235,38 +189,41 @@ define( [
 
          describe( 'and with a configured information resource', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  information: {
+                     resource: 'information'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     information: {
-                        resource: 'information'
-                     }
-                  }
-               };
-               setup( features, setupHttpMockData, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               setupHttpMockData( $httpBackend );
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'publishes a resource with the widget.json content of the selected widget (R2.3)', function() {
                $httpBackend.flush();
-               jasmine.Clock.tick( 0 );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.information', {
                      resource: 'information',
                      data: widgetInformation
@@ -280,38 +237,41 @@ define( [
 
          describe( 'and with a configured bower resource', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  bower: {
+                     resource: 'bower'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     bower: {
-                        resource: 'bower'
-                     }
-                  }
-               };
-               setup( features, setupHttpMockData, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               setupHttpMockData( $httpBackend );
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'publishes a resource with the bower.json content of the selected widget (R2.4)', function() {
                $httpBackend.flush();
-               jasmine.Clock.tick( 0 );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.bower', {
                      resource: 'bower',
                      data: widgetBower
@@ -325,54 +285,47 @@ define( [
 
          describe( 'and with a configured documentation resource and a configured file name list', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  documentation: {
+                     resource: 'documentation',
+                     sources: [ 'README.md', 'SPECIFICATION.md', 'DOCUMENTATION.md' ]
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     documentation: {
-                        resource: 'documentation',
-                        sources: [ 'README.md', 'SPECIFICATION.md', 'DOCUMENTATION.md' ]
-                     }
-                  }
-               };
-               setup( features, function( $injector ) {
-                  $httpBackend = $injector.get( '$httpBackend' );
-                  $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/README.md' )
-                     .respond( 404, {value: 'Not Found'} );
-                  $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/SPECIFICATION.md' )
-                     .respond( 200, {} );
-                  $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/DOCUMENTATION.md' )
-                     .respond( 404, {value: 'Not Found'} );
-               }, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/README.md' )
+                  .respond( 404, {value: 'Not Found'} );
+               $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/SPECIFICATION.md' )
+                  .respond( 200, {} );
+               $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/DOCUMENTATION.md' )
+                  .respond( 404, {value: 'Not Found'} );
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
-            } );
-
-            //////////////////////////////////////////////////////////////////////////////////////////////////
-
-            it( 'tests the file names until one exists (R2.5)', function() {
-               $httpBackend.flush( 2 );
-               $httpBackend.verifyNoOutstandingRequest();
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'publishes a resource with the link to the document of the selected widget (R2.5)', function() {
                $httpBackend.flush();
-               jasmine.Clock.tick( 0 );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.documentation', {
                      resource: 'documentation',
                      data: {
@@ -386,7 +339,7 @@ define( [
                   {
                      deliverToSender: false
                   } );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .not.toHaveBeenCalledWith( 'didEncounterError.HTTP_HEAD', {
                      code: 'HTTP_HEAD',
                      data: {
@@ -401,36 +354,40 @@ define( [
 
          describe( 'and with a configured test runner resource', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  testRunner: {
+                     resource: 'testRunner'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     testRunner: {
-                        resource: 'testRunner'
-                     }
-                  }
-               };
-               setup( features, setupHttpMockData, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               setupHttpMockData( $httpBackend );
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'publishes a resource with the link to the spec_runner.html (R2.6)', function() {
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.testRunner', {
                      resource: 'testRunner',
                      data: {
@@ -448,39 +405,43 @@ define( [
 
          describe( 'and with a configured information resource but the loading process fails', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  information: {
+                     resource: 'information'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     information: {
-                        resource: 'information'
-                     }
-                  }
-               };
-               setup( features, setupHttpMockData404, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               setupHttpMockData404( $httpBackend );
+
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'emits a didEncounterError event with an error message and replaces the resource with an empty object (R2.7)', function() {
                $httpBackend.flush();
-               jasmine.Clock.tick( 0 );
 
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didEncounterError.HTTP_GET', {
                      code: 'HTTP_GET',
                      data: {
@@ -489,7 +450,7 @@ define( [
                      },
                      message: jasmine.any( String )
                   } );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.information', {
                      resource: 'information',
                      data: {}
@@ -503,39 +464,42 @@ define( [
 
          describe( 'and with a configured bower resource but the loading process fails', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  bower: {
+                     resource: 'bower'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     bower: {
-                        resource: 'bower'
-                     }
-                  }
-               };
-               setup( features, setupHttpMockData404, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               setupHttpMockData404( $httpBackend );
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'emits a didEncounterError event with an error message and replaces the resource with an empty object (R2.7)', function() {
                $httpBackend.flush();
-               jasmine.Clock.tick( 0 );
 
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didEncounterError.HTTP_GET', {
                      code: 'HTTP_GET',
                      data: {
@@ -544,7 +508,7 @@ define( [
                      },
                      message: jasmine.any( String )
                   } );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.bower', {
                      resource: 'bower',
                      data: {}
@@ -558,39 +522,43 @@ define( [
 
          describe( 'and with a configured documentation resource but the loading process fails', function() {
 
+            var widgetConfiguration = {
+               list: {
+                  resource: 'listResource'
+               },
+               select: {
+                  parameter: 'widgetId',
+                  documentation: {
+                     resource: 'documentation'
+                  }
+               }
+            };
+
+            createSetup( widgetConfiguration, constructAbsoluteUrl );
+
             beforeEach( function() {
-               var features = {
-                  list: {
-                     resource: 'listResource'
-                  },
-                  select: {
-                     parameter: 'widgetId',
-                     documentation: {
-                        resource: 'documentation'
-                     }
-                  }
-               };
-               setup( features, setupHttpMockData404, constructAbsoluteUrl );
-               testBed.eventBusMock.publish( 'didReplace.listResource', {
+               setupHttpMockData404( $httpBackend );
+
+               axMocks.triggerStartupEvents();
+               testEventBus.publish( 'didReplace.listResource', {
                   resource: 'listResource',
-                  data: specData.listResource
+                  data: data.listResource
                } );
-               jasmine.Clock.tick( 0 );
-               testBed.eventBusMock.publish( 'didNavigate._self', {
+               testEventBus.flush();
+               testEventBus.publish( 'didNavigate._self', {
                   data: {
-                     widgetId: 'AxWidgetBrowserWidget'
+                     widgetId: 'ax-widget-browser-widget'
                   }
                } );
-               jasmine.Clock.tick( 0 );
+               testEventBus.flush();
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'emits a didEncounterError event with an error message and replaces the resource with an empty link (R2.7)', function() {
                $httpBackend.flush();
-               jasmine.Clock.tick( 0 );
 
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didEncounterError.HTTP_GET', {
                      code: 'HTTP_GET',
                      data: {
@@ -598,7 +566,7 @@ define( [
                      },
                      message: jasmine.any( String )
                   } );
-               expect( testBed.scope.eventBus.publish )
+               expect( widgetEventBus.publish )
                   .toHaveBeenCalledWith( 'didReplace.documentation', {
                      resource: 'documentation',
                      data: {
@@ -615,5 +583,59 @@ define( [
             } );
          } );
       } );
+
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+      function setupHttpMockData( $httpBackend ) {
+         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json' )
+            .respond( 200, widgetInformation );
+         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/bower.json' )
+            .respond( 200, widgetBower );
+         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/SPECIFICATION.md' )
+            .respond( 200, {} );
+         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/README.md' )
+            .respond( 404, { value: 'Not Found' } );
+         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/DOCUMENTATION.md' )
+            .respond( 404, { value: 'Not Found' } );
+      }
+
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+      function setupHttpMockData404( $httpBackend ) {
+         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json' )
+            .respond( 404, { value: 'Not Found' } );
+         $httpBackend.when( 'GET', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/bower.json' )
+            .respond( 404, { value: 'Not Found' } );
+         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/SPECIFICATION.md' )
+            .respond( 404, { value: 'Not Found' } );
+         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/README.md' )
+            .respond( 404, { value: 'Not Found' } );
+         $httpBackend.when( 'HEAD', 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/DOCUMENTATION.md' )
+            .respond( 404, { value: 'Not Found' } );
+      }
+
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+      function constructAbsoluteUrl( place, optionalParameters ) {
+         return 'http://localhost:8000/index.html#/widgetBrowser/' +
+                optionalParameters[ widgetScope.features.select.parameter ];
+      }
+
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+      function constructAbsoluteUrlWithAdditionalParameters( place, optionalParameters ) {
+         var additionalParameters = '';
+         if( optionalParameters.anchor === undefined ) {
+            additionalParameters = additionalParameters + '/references';
+         }
+         if( optionalParameters.userId === undefined ) {
+            additionalParameters = additionalParameters + '/03804';
+         }
+         if( optionalParameters.processId === undefined ) {
+            additionalParameters = additionalParameters + '/73929';
+         }
+         return 'http://localhost:8000/index.html#/widgetBrowser/' +
+                optionalParameters[ widgetScope.features.select.parameter ] + additionalParameters;
+      }
    } );
 } );