+++ includes/widgets/widget-browser/ax-widget-browser-widget/spec/data/fixtures.js
@@ -1,0 +1,133 @@
+/**
+ * Copyright 2015 aixigo AG
+ * Released under the MIT license.
+ * http://laxarjs.org/license
+ */
+define( {
+   'listResource': {
+      widgets: [
+         {
+            name: 'ax-headline-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-headline-widget/widget.json'
+         },
+         {
+            name: 'ax-markdown-display-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-markdown-display-widget/widget.json'
+         },
+         {
+            name: 'ax-media-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-media-widget/widget.json'
+         },
+         {
+            name: 'ax-messages-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-messages-widget/widget.json'
+         },
+         {
+            name: 'ax-widget-browser-widget',
+            specification: 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json'
+         },
+         {
+            name: 'ax-widget-information-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-widget-information-widget/widget.json'
+         },
+         {
+            name: 'ax-widget-listing-activity',
+            specification: 'myIncludes/widgets/laxarjs/ax-widget-listing-activity/widget.json'
+         }
+      ]
+   },
+   'browserList': [
+      {
+         name: 'ax-headline-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-headline-widget',
+         specification: 'myIncludes/widgets/laxarjs/ax-headline-widget/widget.json'
+      },
+      {
+         name: 'ax-markdown-display-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-markdown-display-widget',
+         specification: 'myIncludes/widgets/laxarjs/ax-markdown-display-widget/widget.json'
+      },
+      {
+         name: 'ax-media-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-media-widget',
+         specification: 'myIncludes/widgets/laxarjs/ax-media-widget/widget.json'
+      },
+      {
+         name: 'ax-messages-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-messages-widget',
+         specification: 'myIncludes/widgets/laxarjs/ax-messages-widget/widget.json'
+      },
+      {
+         name: 'ax-widget-browser-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-widget-browser-widget',
+         specification: 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json'
+      },
+      {
+         name: 'ax-widget-information-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-widget-information-widget',
+         specification: 'myIncludes/widgets/laxarjs/ax-widget-information-widget/widget.json'
+      },
+      {
+         name: 'ax-widget-listing-activity',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-widget-listing-activity',
+         specification: 'myIncludes/widgets/laxarjs/ax-widget-listing-activity/widget.json'
+      }
+   ],
+   'browserListWithParameters': [
+      {
+         name: 'ax-headline-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-headline-widget/73929',
+         specification: 'myIncludes/widgets/laxarjs/ax-headline-widget/widget.json'
+      },
+      {
+         name: 'ax-markdown-display-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-markdown-display-widget/73929',
+         specification: 'myIncludes/widgets/laxarjs/ax-markdown-display-widget/widget.json'
+      },
+      {
+         name: 'ax-media-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-media-widget/73929',
+         specification: 'myIncludes/widgets/laxarjs/ax-media-widget/widget.json'
+      },
+      {
+         name: 'ax-messages-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-messages-widget/73929',
+         specification: 'myIncludes/widgets/laxarjs/ax-messages-widget/widget.json'
+      },
+      {
+         name: 'ax-widget-browser-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-widget-browser-widget/73929',
+         specification: 'myIncludes/widgets/widget-browser/ax-widget-browser-widget/widget.json'
+      },
+      {
+         name: 'ax-widget-information-widget',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-widget-information-widget/73929',
+         specification: 'myIncludes/widgets/laxarjs/ax-widget-information-widget/widget.json'
+      },
+      {
+         name: 'ax-widget-listing-activity',
+         href: 'http://localhost:8000/index.html#/widgetBrowser/ax-widget-listing-activity/73929',
+         specification: 'myIncludes/widgets/laxarjs/ax-widget-listing-activity/widget.json'
+      }
+   ],
+   'smallerListResource': {
+      widgets: [
+         {
+            name: 'ax-headline-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-headline-widget/widget.json'
+         },
+         {
+            name: 'ax-messages-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-messages-widget/widget.json'
+         },
+         {
+            name: 'ax-widget-information-widget',
+            specification: 'myIncludes/widgets/laxarjs/ax-widget-information-widget/widget.json'
+         },
+         {
+            name: 'ax-widget-listing-activity',
+            specification: 'myIncludes/widgets/laxarjs/ax-widget-listing-activity/widget.json'
+         }
+      ]
+   }
+} );