+++ includes/widgets/widget-browser/ax-widget-listing-activity/spec/spec_runner.js
@@ -1,15 +1,17 @@
 /**
  * Copyright 2015 aixigo AG
  * Released under the MIT license.
- * http://laxarjs.org/license
+ * http://www.laxarjs.org
  */
 ( function( global ) {
    'use strict';
    global.laxarSpec = {
-      title: 'AxWidgetListingActivity Specification',
+      title: 'ax-widget-listing-activity Specification',
       tests: [
          'ax-widget-listing-activity.spec'
       ],
-      requireConfig: {}
+      testRunner: 'laxar-mocks',
+      jasmineMajorVersion: 2
    };
 } )( this );
+